Added delete playlist action
