+++ src-tauri/src/operations/playlist_operations.rs
@@ -2,6 +2,7 @@ use diesel::prelude::*;
 use crate::models::playlist_model:: {
     NewPlaylist, Playlist, PlaylistArg 
 };
+use crate::models::music_model::Music;
 use crate::models::playlist_music_model::NewPlaylistMusic;
 use crate::db::establish_connection;
 
@@ -117,5 +118,51 @@ pub fn insert_song_into_playlist(playlist_id_arg: i32, music_id_arg: i32) -> Res
       Err(err) => {
           Err(format!("Error: {}", err))
       }
+  }
+}
+
+#[tauri::command]
+pub fn get_all_music_from_playlist(playlist_id_arg: i32) -> Result<Vec<Music>, String> {
+  use crate::schema::playlist_music::dsl::*;
+  use crate::schema::music::dsl::*;
+
+  let mut connection: SqliteConnection = establish_connection();
+
+  let music_id_list: Vec<i32> = match playlist_music
+  .filter(playlist_id.eq(playlist_id_arg))
+  .select(music_id) 
+  .load::<i32>(&mut connection) {
+      Ok(ids) => ids,
+      Err(err) => {
+        eprintln!("Error loading music IDs from playlist: {}", err);
+        return Err(format!("Error querying playlist music entries: {}", err));
+      }
+  };
+
+  let music_list: Vec<Music> = match music
+  .filter(id.eq_any(music_id_list)) // Filter by the list of music_ids
+  .load::<Music>(&mut connection) {
+      Ok(result) => result,
+      Err(err) => {
+          eprintln!("Error loading music: {}", err);
+          return Err(format!("Error querying music: {}", err));
+      }
+  };
+
+  Ok(music_list)
+}
+
+#[tauri::command]
+pub fn delete_playlist(playlist_id_arg: i32) -> Result<(), String> {
+  use crate::schema::playlist::dsl::*;
+
+  let mut connection: SqliteConnection = establish_connection();
+
+  let result: Result<usize, diesel::result::Error> = diesel::delete(playlist.filter(id.eq(playlist_id_arg)))
+    .execute(&mut connection);
+
+  match result {
+    Ok(_) => Ok(()),
+    Err(err) => Err(format!("Error deleting playlist entry: {}", err)), // Return error to the client
   }
 }
